+++ src/powernse/reading/__init__.py
@@ -2,6 +2,9 @@
 
 from powernse.reading.bhavcopy import BhavcopyReader
 from powernse.reading.corporate import CorporateActionReader
+from powernse.reading.coverage import CoverageReader
+from powernse.reading.deals import DealsReader, SecbanReader
+from powernse.reading.delivery import DeliveryReader
 from powernse.reading.futures import FoReader
 from powernse.reading.index import IndexReader
 from powernse.reading.snapshots import SnapshotReader
@@ -9,7 +12,11 @@ from powernse.reading.snapshots import SnapshotReader
 __all__ = [
     "BhavcopyReader",
     "CorporateActionReader",
+    "CoverageReader",
+    "DealsReader",
+    "DeliveryReader",
     "FoReader",
     "IndexReader",
+    "SecbanReader",
     "SnapshotReader",
 ]